Replace old-Prof with Prof2
As of this commit:
 - src/tool/hpcprof{,-mpi} is the Prof2 frontend code,
 - src/lib/profile is the Prof2 backend code (no change),
 - the installed hpcprof{,-mpi} is Prof2, and
 - `make check` passes.

Removed:
 - Old-Prof frontend code in src/tool/hpcprof{,-mpi}. Part of the Git history
   so can be inspected if needed.
 - src/tool/hpcprof2{,-mpi}. Leftover changes should be rebased
   onto src/tool/hpcprof{,-mpi} (Git *should* do this automatically, I think).
 - hpcprof2{,-mpi} is no longer installed. Use hpcprof{,-mpi} instead.
