+++ src/tool/Makefile.am
@@ -71,7 +71,6 @@ if OPT_BUILD_TOOL_ALL
 SUBDIRS += \
 	hpcstruct \
 	hpcprof \
-	hpcprof2 \
 	hpcproftt \
 	hpclump \
 	hpctracedump
@@ -89,7 +88,6 @@ endif
 
 if OPT_ENABLE_MPI
   SUBDIRS += hpcprof-mpi
-  SUBDIRS += hpcprof2-mpi
 
 if OPT_ENABLE_HPCSERVER
   SUBDIRS += hpcserver/mpi
