+++ src/tool/hpcprof-mpi/master.cpp
@@ -46,19 +46,26 @@
 
 #include "lib/profile/util/vgannotations.hpp"
 
-#include "args.hpp"
+#include "tree.hpp"
+#include "../hpcprof/args.hpp"
 
 #include "lib/profile/pipeline.hpp"
+#include "lib/profile/packedids.hpp"
 #include "lib/profile/source.hpp"
+#include "lib/profile/sources/packed.hpp"
 #include "lib/profile/sinks/experimentxml4.hpp"
 #include "lib/profile/sinks/hpctracedb2.hpp"
 #include "lib/profile/sinks/sparsedb.hpp"
 #include "lib/profile/finalizers/denseids.hpp"
 #include "lib/profile/finalizers/directclassification.hpp"
+#include "lib/profile/util/log.hpp"
+#include "lib/profile/mpi/all.hpp"
 
 #include <iostream>
+#include <stack>
 
 using namespace hpctoolkit;
+using namespace hpctoolkit::literals;
 namespace fs = stdshim::filesystem;
 
 template<class T, class... Args>
@@ -66,23 +73,26 @@ static std::unique_ptr<T> make_unique_x(Args&&... args) {
   return std::unique_ptr<T>(new T(std::forward<Args>(args)...));
 }
 
-int main(int argc, char* const argv[]) {
-  // Read in the arguments.
-  ProfArgs args(argc, argv);
-
-  // Get the main core of the Pipeline set up.
+int rank0(ProfArgs&& args) {
+  // We only have one Pipeline, this is its builder.
   ProfilePipeline::Settings pipelineB;
-  for(auto& sp : args.sources) pipelineB << std::move(sp.first);
-  for(auto& sp : args.ksyms) pipelineB << std::move(sp);
-  for(auto& sp : args.structs) pipelineB << std::move(sp.first);
-  ProfArgs::StructWarner sw(args);
-  pipelineB << sw;
+  for(auto& sp: args.sources) pipelineB << std::move(sp.first);
+  for(auto& sp: args.ksyms) pipelineB << std::move(sp);
+
+  // Set up our reduction tree with our fellow peers
+  RankTree tree{std::max<std::size_t>(args.threads, 2)};
+
+  // When time, gather up the data from all our friends and emit it.
+  Receiver::append(pipelineB, tree);
+
+  // Make sure all Metrics have their Statistics set properly
   ProfArgs::StatisticsExtender se(args);
   pipelineB << se;
 
-  // Provide Ids for things from the void
-  finalizers::DenseIds dids;
-  pipelineB << dids;
+  // Load in the Finalizers for Structfiles.
+  for(auto& sp: args.structs) pipelineB << std::move(sp.first);
+  ProfArgs::StructWarner sw(args);
+  pipelineB << sw;
 
   // Make sure the files are searched for as they should be
   ProfArgs::Prefixer pr(args);
@@ -93,6 +103,65 @@ int main(int argc, char* const argv[]) {
   finalizers::DirectClassification dc(args.dwarfMaxSize);
   pipelineB << dc;
 
+  // Ids for everything are pulled from the void. We call the shots here.
+  finalizers::DenseIds dids;
+  pipelineB << dids;
+
+  // Ensure that our Thread ids are unique with respect to the workers.
+  // Since this is rank 0, we don't need to adjust our ids.
+  struct ThreadIDUniquer : public ProfileSink {
+    void write() override {}
+    DataClass accepts() const noexcept override { return DataClass::threads; }
+    ExtensionClass requires() const noexcept override { return {}; }
+    DataClass wavefronts() const noexcept override { return DataClass::threads; }
+    void notifyPipeline() noexcept override {
+      src.registerOrderedWavefront();
+    }
+    void notifyWavefront(DataClass wave) override {
+      if(!wave.hasThreads()) return;
+      auto mpiSem = src.enterOrderedWavefront();
+      mpi::exscan(src.threads().size(), mpi::Op::sum());
+    }
+  } tiduniquer;
+  pipelineB << tiduniquer;
+
+  // No-op Source to detect whether the Pipeline requires timepoints
+  // It should be args.include_traces, but you can never be too certain
+  struct Detector : public ProfileSource {
+    DataClass provides() const noexcept override { return DataClass::ctxTimepoints; }
+    DataClass finalizeRequest(const DataClass& d) const noexcept override { return d; }
+    void read(const DataClass&) override {};
+    bool needsTimepoints() const { return sink.limit().hasCtxTimepoints(); }
+  } detector;
+  pipelineB << detector;
+
+  // When everything is ready, ship off the block to the workers.
+  struct Sender : public IdPacker {
+    Sender(Detector& d) : detector(d) {};
+    void notifyPipeline() noexcept override {
+      IdPacker::notifyPipeline();
+      src.registerOrderedWavefront();
+    }
+    void notifyPacked(std::vector<uint8_t>&& block) override {
+      auto mpiSem = src.enterOrderedWavefront();
+      mpi::bcast(std::move(block), 0);
+      mpi::bcast((uint8_t)(detector.needsTimepoints() ? 1 : 0), 0);
+    }
+    Detector& detector;
+  } spacker(detector);
+  pipelineB << spacker;
+
+  // For unpacking metrics, we need to be able to map the IDs back to their
+  // Contexts. This does the magic for us.
+  sources::Packed::ctx_map_t cmap;
+  sources::Packed::ContextTracker ctracker(cmap);
+  pipelineB << ctracker;
+
+  // The workers will come back to us with the metric data, so we have to be
+  // ready to accept it.
+  MetricReceiver::append(pipelineB, tree, cmap);
+
+  // Finally, eventually we get to actually write stuff out.
   switch(args.format) {
   case ProfArgs::Format::sparse: {
     std::unique_ptr<sinks::HPCTraceDB2> tdb;
@@ -106,13 +175,14 @@ int main(int argc, char* const argv[]) {
   }
   }
 
-  // Create the Pipeline, let the fun begin.
+  // Create and drain the Pipeline, that's all we do.
   ProfilePipeline pipeline(std::move(pipelineB), args.threads);
-
-  // Drain the Pipeline, and make everything happen.
   pipeline.run();
 
-  if(args.valgrindUnclean) std::exit(0);  // Skips local cleanup of pipeline
+  if(args.valgrindUnclean) {
+    mpi::World::finalize();
+    std::exit(0);
+  }
 
   return 0;
 }
