--- src/tool/hpcprof/Args.cpp
@@ -1,126 +1,0 @@
-// -*-Mode: C++;-*-
-
-// * BeginRiceCopyright *****************************************************
-//
-// $HeadURL$
-// $Id$
-//
-// --------------------------------------------------------------------------
-// Part of HPCToolkit (hpctoolkit.org)
-//
-// Information about sources of support for research and development of
-// HPCToolkit is at 'hpctoolkit.org' and in 'README.Acknowledgments'.
-// --------------------------------------------------------------------------
-//
-// Copyright ((c)) 2002-2022, Rice University
-// All rights reserved.
-//
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions are
-// met:
-//
-// * Redistributions of source code must retain the above copyright
-//   notice, this list of conditions and the following disclaimer.
-//
-// * Redistributions in binary form must reproduce the above copyright
-//   notice, this list of conditions and the following disclaimer in the
-//   documentation and/or other materials provided with the distribution.
-//
-// * Neither the name of Rice University (RICE) nor the names of its
-//   contributors may be used to endorse or promote products derived from
-//   this software without specific prior written permission.
-//
-// This software is provided by RICE and contributors "as is" and any
-// express or implied warranties, including, but not limited to, the
-// implied warranties of merchantability and fitness for a particular
-// purpose are disclaimed. In no event shall RICE or contributors be
-// liable for any direct, indirect, incidental, special, exemplary, or
-// consequential damages (including, but not limited to, procurement of
-// substitute goods or services; loss of use, data, or profits; or
-// business interruption) however caused and on any theory of liability,
-// whether in contract, strict liability, or tort (including negligence
-// or otherwise) arising in any way out of the use of this software, even
-// if advised of the possibility of such damage.
-//
-// ******************************************************* EndRiceCopyright *
-
-//***************************************************************************
-//
-// File:
-//   $HeadURL$
-//
-// Purpose:
-//   [The purpose of this file]
-//
-// Description:
-//   [The set of functions, macros, etc. defined in the file]
-//
-//***************************************************************************
-
-//************************* System Include Files ****************************
-
-#include <iostream>
-using std::cerr;
-using std::endl;
-
-#include <string>
-using std::string;
-
-//*************************** User Include Files ****************************
-
-#include "Args.hpp"
-
-//*************************** Forward Declarations **************************
-
-// Cf. DIAG_Die.
-#define ARG_ERROR(streamArgs)                                        \
-  { std::ostringstream WeIrDnAmE;                                    \
-    WeIrDnAmE << streamArgs /*<< std::ends*/;                        \
-    printError(std::cerr, WeIrDnAmE.str());                          \
-    exit(1); }
-
-//***************************************************************************
-
-
-//***************************************************************************
-// Args
-//***************************************************************************
-
-Args::Args()
-{
-  hpcprof_isMetricArg = false;
-  hpcprof_forceMetrics = false;
-}
-
-
-Args::~Args()
-{
-}
-
-
-void
-Args::parse(int argc, const char* const argv[])
-{
-  ArgsHPCProf::parse(argc, argv, Analysis::AppType::APP_HPCPROF);
-
-  if (parser.isOpt("metric")) {
-    hpcprof_isMetricArg = true;
-  }
-
-  if (parser.isOpt("force-metric")) {
-    hpcprof_forceMetrics = true;
-  }
-
-  // Currently, hpcprof does not generate thread-level metric db
-  db_makeMetricDB = false;
-}
-
-
-const std::string
-Args::getCmd() const
-{
-  // avoid error messages with: <path>/hpcprof-mpi
-  static string cmd = "hpcprof";
-  return cmd; // parser.getCmd();
-}
-
